Refactor and improve chat interface
This commit refactors the chat message component into smaller, more manageable sub-components. It also improves the visual design of the chat messages and adds error handling for message sending.

The `Message` component has been broken down into:
- `MessageHeader`: Displays the sender's avatar and name.
- `MessageContent`: Renders the message content, including markdown and file previews.
- `MessageActionsBar`: Renders the action buttons for a message.

The visual design has been improved by:
- Adjusting spacing and padding for a cleaner look.
- Making the background color for assistant messages more subtle.
- Removing the redundant assistant name from each message.

Error handling has been added to display an error message when a message fails to send. This is achieved by adding a `status` field to the `ChatMessage` type and updating the UI to reflect the message status.
