+++ app/[locale]/globals.css
@@ -48,7 +48,7 @@
     --primary: 0 0% 9%;
     --primary-foreground: 0 0% 98%;
 
-    --secondary: 0 0% 96.1%;
+    --secondary: 0 0% 98%;
     --secondary-foreground: 0 0% 9%;
 
     --accent: 0 0% 96.1%;
