+++ components/chat/chat-messages.tsx
@@ -3,6 +3,7 @@ import { ChatbotUIContext } from "@/context/context"
 import { Tables } from "@/supabase/types"
 import { FC, useContext, useState } from "react"
 import { Message } from "../messages/message"
+import { ChatMessage } from "@/types"
 
 interface ChatMessagesProps {}
 
@@ -11,7 +12,7 @@ export const ChatMessages: FC<ChatMessagesProps> = ({}) => {
 
   const { handleSendEdit } = useChatHandler()
 
-  const [editingMessage, setEditingMessage] = useState<Tables<"messages">>()
+  const [editingMessage, setEditingMessage] = useState<ChatMessage>()
 
   return chatMessages
     .sort((a, b) => a.message.sequence_number - b.message.sequence_number)
@@ -25,9 +26,9 @@ export const ChatMessages: FC<ChatMessagesProps> = ({}) => {
       return (
         <Message
           key={chatMessage.message.sequence_number}
-          message={chatMessage.message}
+          chatMessage={chatMessage}
           fileItems={messageFileItems}
-          isEditing={editingMessage?.id === chatMessage.message.id}
+          isEditing={editingMessage?.message.id === chatMessage.message.id}
           isLast={index === array.length - 1}
           onStartEdit={setEditingMessage}
           onCancelEdit={() => setEditingMessage(undefined)}
